refactor(api): rename GiantBombApi to RawgApi and complete migration
